Extract ARCoreDependenciesHelper class to simplify MainActivity
